Move tests to own module-test file
